+++ spring/src/main/java/ru/yoomoney/tech/dbqueue/spring/dao/Oracle11QueueDao.java
@@ -1,5 +1,11 @@
 package ru.yoomoney.tech.dbqueue.spring.dao;
 
+import java.time.Duration;
+import java.util.List;
+import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.stream.Collectors;
+import javax.annotation.Nonnull;
 import org.springframework.jdbc.core.JdbcOperations;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
@@ -7,12 +13,6 @@ import ru.yoomoney.tech.dbqueue.api.EnqueueParams;
 import ru.yoomoney.tech.dbqueue.config.QueueTableSchema;
 import ru.yoomoney.tech.dbqueue.dao.QueueDao;
 import ru.yoomoney.tech.dbqueue.settings.QueueLocation;
-
-import javax.annotation.Nonnull;
-import java.time.Duration;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.stream.Collectors;
 
 import static java.util.Objects.requireNonNull;
 
@@ -69,8 +69,12 @@ public class Oracle11QueueDao implements QueueDao {
         jdbcTemplate.update(enqueueSqlCache.computeIfAbsent(location, this::createEnqueueSql), params);
         return generatedId;
     }
-
-
+    
+    @Override
+    public void enqueueBatch(@Nonnull QueueLocation location, @Nonnull List<EnqueueParams<String>> enqueueParams) {
+        throw new UnsupportedOperationException("batch enqueue is not supported for this type of database");
+    }
+    
     @Override
     public boolean deleteTask(@Nonnull QueueLocation location, long taskId) {
         requireNonNull(location);