+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-### NEXT_VERSION_TYPE=MAJOR|MINOR|PATCH
+### NEXT_VERSION_TYPE=MAJOR
 ### NEXT_VERSION_DESCRIPTION_BEGIN
+* Batch enqueue feature was implemented for postgres
+**breaking changes** New methods in interfaces: `QueueDao`, `QueueProducer`.
 ### NEXT_VERSION_DESCRIPTION_END
 ## [16.2.0](https://bitbucket.yooteam.ru/projects/BACKEND-LIBRARIES/repos/db-queue/pull-requests/8) (30-10-2023)